EuroMilRegister/requirements.txt
grpcio==1.60.0; python_version >= '3.7'
grpcio-tools==1.60.0; python_version >= '3.7'